CVSL-1394 refactored file names and moved to appropriate package. Also added error messages
