+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/policies/Codes.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.policies
 
 const val ELECTRONIC_TAG_COND_CODE = "fd129172-bdd3-4d97-a4a0-efd7b47a49d4"
 const val EXCLUSION_ZONE_COND_CODE = "0f9a20f4-35c7-4c77-8af8-f200f153fa11"
